rpi/requirements.txt
AWSIoTPythonSDK==1.5.2
RPi.GPIO==0.7.1
